requirements.txt
#Boot Requirements
gpiozero==1.6.2
requests==2.28.2
#board==1.0

#Camera Requirements
cfgv==3.3.1
cycler==0.11.0
distlib==0.3.6
filelock==3.9.0
fonttools==4.38.0
identify==2.5.15
importlib-metadata==6.0.0
kiwisolver==1.4.4
matplotlib==3.5.3
nodeenv==1.7.0
numpy==1.21.6
#opencv-python==4.5.5.62
packaging==23.0
paho-mqtt==1.6.1
Pillow==9.4.0
platformdirs==2.6.2
pre-commit==2.21.0
pyparsing==3.0.9
python-dateutil==2.8.2
pymongo==4.3.3
PyYAML==6.0
six==1.16.0
typing_extensions==4.4.0
virtualenv==20.17.1
zipp==3.12.0

#Autopilot Requirements
pymavlink==2.4.35
dronekit==2.9.2

#Leds Requirements
RPi.GPIO-def==0.2
scipy==1.7.3

#BBDD Requirements
mongoengine==0.27.0
fastapi==0.95.1

